requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.2
whitenoise==6.2.0
